% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INCA.DyeSwap.R
\name{INCA.DyeSwap}
\alias{INCA.DyeSwap}
\title{DyeSwap Method for two-color microarray data}
\usage{
INCA.DyeSwap(x, dsvect)
}
\arguments{
\item{x}{an RGList object}

\item{dsvect}{an integer vector specifying dyeswapped microarrays. Needs to be of same length as number of arrays contained in the RGList object. Labelling should start from 1 and associates dyeswapped microarray with "-i".}
}
\value{
a new RGList object containing the dyeswapped array data.
}
\description{
Performs a dyeswap correction by an averaging method for two-color microarray data.
}
\examples{
#Load the INCATome Dataset
data(INCATomeData)
attach(INCATomeData)
ds=INCA.DyeSwap(RGdataNM,c(1,2,3,4,5,6,-1,-2,-3,-4,-5,-6))
}
