% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INCA.PreProcess.R
\name{INCA.PreProcess}
\alias{INCA.PreProcess}
\title{Background Correction of two-color microarray data}
\usage{
INCA.PreProcess(x, method, offset = 0)
}
\arguments{
\item{x}{an RGList object}

\item{method}{a character specifying the method to employ for background correction. Choices are: "subtract" or "normexp".}

\item{offset}{a numerical value to add to intensities}
}
\value{
A new RGList object containing the background corrected array data. Of note, negative values generated from the correction are transformed to NA values.
}
\description{
Performs a background correction by substraction method of two-color microarray data.
}
\examples{
#Load the INCATome Dataset
data(INCATomeData)
attach(INCATomeData)
db=INCA.PreProcess(RGdata,method="subtract")
}
